+++ src/core/providers/ContextProvider.tsx
@@ -2,17 +2,18 @@
 
 import { useState } from 'react';
 import PokemonListContext from '../context/pokemonListContext';
-import { PokemonDataType } from '@/pages/Home/types';
-import { pokemonDataDefault } from '@/pages/Home/data/pokemonDataDefault';
+import pokemonDataDefault from '@/pages/Home/data/pokemonDataDefault';
 import DeckListContext from '../context/deckListContext';
+import { PokemonDataType } from '../api/types';
 
 export default function ContextProvider({
   children
 }: {
   children: React.ReactNode;
 }) {
-  const [pokemonData, setPokemonData] =
-    useState<PokemonDataType[]>(pokemonDataDefault);
+  const [pokemonData, setPokemonData] = useState<PokemonDataType[]>(
+    pokemonDataDefault()
+  );
   const [deckList, setDeckList] = useState<number[]>([]);
 
   return (
