+++ src/pages/Home/constant/constant.ts
@@ -1,1 +1,2 @@
-export const limitStartSearching = 2;
+const limitStartSearching = (): number => 2;
+export default limitStartSearching;
